+++ changelog.md
@@ -12,6 +12,7 @@ This version is not yet released. If you are reading this on the website, then t
 - Make sided [`fill ⬚`](https://uiua.org/docs/fill) work with [`stencil ⧈`](https://uiua.org/docs/stencil)
 - Add sided subscripts for [`join ⊂`](https://uiua.org/docs/join), which joins with one side as the list
 - [`anti ⌝`](https://uiua.org/docs/anti) [`keep ▽`](https://uiua.org/docs/keep) now accepts non-boolean integers
+- Add [`compress`](https://uiua.org/docs/compress) function to compress arrays of bytes
 - Change [`reach ∪`](https://uiua.org/docs/reach)'s glyph
 - Change how sided [`reach ∪`](https://uiua.org/docs/reach) works
 - Add [`&camlist`](https://uiua.org/docs/&camlist) function to list available webcams